backend/requirements.txt
fastapi==0.97.0
pydantic==1.10.9
uvicorn==0.22.0
sqlalchemy==2.0.16
pymysql==1.0.3
cryptography==40.0.2
alembic==1.11.1
pyJWT==2.7.0
pytz==2023.3
passlib[bcrypt]
python-multipart
firebase_admin